requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.31
botocore==1.37.31
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.74.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
s3transfer==0.11.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1

# Added dependencies for Dia
torch>=2.0.0
soundfile>=0.12.1
transformers>=4.36.0
librosa>=0.10.0
scipy>=1.11.0
numpy>=1.24.0